Agrego validaciones del registro y login

+++ src/controllers/apiController.js
@@ -5,6 +5,7 @@ const jwt = require('jsonwebtoken')
 const expressJwt = require('express-jwt')
 const LoginFails = require("../models/LoginFails")
 const fs = require('fs')
+const { validationResult } = require('express-validator')
 
 //const validateToken = expressJwt({ "secret": 'mi-secreto', "algorithm": ['HS256'] })
 const signToken = userId => jwt.sign({ userId }, 'mi-secreto')
@@ -12,28 +13,34 @@ const signToken = userId => jwt.sign({ userId }, 'mi-secreto')
 const apiController = {
     createUser: async (req, res) => {
         try {
+            let errors = validationResult(req)
             const userToCreate = { ...req.body }
             const user = await Users.findAll()
             const isUser = await user.filter(x => x.email === userToCreate.email)
-            if (isUser == "") {
-                const newUser = {
-                    name: userToCreate.name,
-                    lastName: userToCreate.lastName,
-                    userName: userToCreate.userName,
-                    email: userToCreate.email,
-                    password: bcryptjs.hashSync(req.body.password, 10),
-                    roleId: userToCreate.roleId,
-                    accountStateId: userToCreate.accountStateId,
-                    userLocked: false,
-                    createdDate: new Date()
+            if (errors.isEmpty()) {
+                if (isUser == "") {
+                    const newUser = {
+                        name: userToCreate.name,
+                        lastName: userToCreate.lastName,
+                        userName: userToCreate.userName,
+                        email: userToCreate.email,
+                        password: bcryptjs.hashSync(req.body.password, 10),
+                        roleId: userToCreate.roleId,
+                        accountStateId: userToCreate.accountStateId,
+                        userLocked: false,
+                        createdDate: new Date()
+                    }
+                    const createdUser = Users.create(newUser)
+                    const signed = signToken(createdUser.userId)
+
+
+                    return res.status(200).send(signed)
+                } else {
+                    return res.status(403).send('Ese usuario ya se encuentra registrado')
                 }
-                const createdUser = Users.create(newUser)
-                const signed = signToken(createdUser.userId)
-
-
-                return res.status(200).send(signed)
             } else {
-                return res.status(403).send('Ese usuario ya se encuentra registrado')
+                console.log(errors.array())
+                return res.status(403).send('debes completar los campos requeridos')
             }
         }
         catch (err) {
@@ -43,40 +50,46 @@ const apiController = {
     },
     loginUser: async (req, res) => {
         try {
+            const errors = validationResult(req)
             const userToLogin = { ...req.body }
             const user = await Users.findAll()
             const isUser = await user.filter(x => x.email === userToLogin.email)
-
-            if (isUser == "") {
-                res.status(404).send('no se encontro')
-            } else {
-                if (isUser[0].accountStateId === 1) {
-                    const match = await bcryptjs.compareSync(userToLogin.password, isUser[0].password)
-                    if (!match) {
-                        const userId = isUser[0].userId
-                        const userFails = await LoginFails.findAll({ where: { userId: userId } })
-                        if (userFails === " " || userFails.length < 5) {
-                            await LoginFails.create({ userId: userId, failedDate: new Date() })
-                            return res.status(403).send('credenciales incorrectas ' + match)
+            if (errors.isEmpty()) {
+                if (isUser == "") {
+                    res.status(404).json({ mensaje: 'Usuario es inconrrecto' })
+                } else {
+                    if (isUser[0].accountStateId === 1) {
+                        const match = await bcryptjs.compareSync(userToLogin.password, isUser[0].password)
+                        if (!match) {
+                            const userId = isUser[0].userId
+                            const userFails = await LoginFails.findAll({ where: { userId: userId } })
+                            if (userFails === " " || userFails.length < 5) {
+                                await LoginFails.create({ userId: userId, failedDate: new Date() })
+                                return res.status(403).json({ mensaje: 'credenciales incorrectas ' })
+                            } else {
+                                //Acá habráa que bloquear el estado del usuario
+                                await Users.update({ accountStateId: 2 }, { where: { userId: userId } })
+                                res.json(userFails)
+                            }
                         } else {
-                            //Acá habráa que bloquear el estado del usuario
-                            await Users.update({ accountStateId: 2 }, { where: { userId: userId } })
-                            res.json(userFails)
+                            const userId = isUser[0].userId
+                            await LoginFails.destroy({ where: { userId: userId } })
+                            const signed = signToken(userId)
+                            req.session.userLogged = isUser[0]
+                            const log = req.session.userLogged = isUser[0]
+                            console.log(log)
+                            return res.status(200).json({ signed: signed })
+                            //Acá se podría agregar la parte de session y el token firmado
                         }
                     } else {
-                        const userId = isUser[0].userId
-                        await LoginFails.destroy({ where: { userId: userId } })
-                        const signed = signToken(userId)
-                        req.session.userLogged = isUser[0]
-                        const log = req.session.userLogged = isUser[0]
-                        console.log(log)
-                        return res.status(200).send(signed)
-                        //Acá se podría agregar la parte de session y el token firmado
+                        return res.status(403).json({ mensaje: 'Contactese con soporte' })
                     }
-                } else {
-                    return res.status(403).send('Contactese con soporte')
                 }
+            } else {
+                console.log(errors.array())
+                return res.status(403).send('debes completar los campos requeridos')
             }
+
         } catch (err) {
             res.json(err.message)
         }
